Move pyiceberg directory up, remove src/ directory
