+++ python/Makefile
@@ -23,7 +23,7 @@ lint:
 	poetry run pre-commit run --all-files
 
 test:
-	poetry run coverage run --source=src/ -m pytest tests/
+	poetry run coverage run --source=pyiceberg/ -m pytest tests/
 	poetry run coverage report -m --fail-under=90
 	poetry run coverage html
 	poetry run coverage xml
