+++ python/pyiceberg/avro/codecs/bzip2.py
