+++ python/pyiceberg/avro/codecs/snappy_codec.py
