+++ python/pyiceberg/avro/file.py
