+++ python/pyiceberg/avro/resolver.py
