+++ python/pyiceberg/catalog/__init__.py
