+++ python/pyiceberg/catalog/base.py
