+++ python/pyiceberg/exceptions.py
