+++ python/pyiceberg/expressions/base.py
