+++ python/pyiceberg/io/base.py
