+++ python/pyiceberg/io/memory.py
