+++ python/pyiceberg/schema.py
