+++ python/pyiceberg/serializers.py
