+++ python/pyiceberg/table/__init__.py
