+++ python/pyiceberg/table/base.py
