+++ python/pyiceberg/table/metadata.py
