+++ python/pyiceberg/table/partitioning.py
