+++ python/pyiceberg/utils/bin_packing.py
