+++ python/pyiceberg/utils/iceberg_base_model.py
