+++ python/pyiceberg/utils/singleton.py
