+++ python/pyproject.toml
@@ -34,7 +34,7 @@ classifiers = [
 ]
 
 packages = [
-    { include = "pyiceberg", from = "src" },
+    { include = "pyiceberg" },
 ]
 
 [tool.poetry.dependencies]
@@ -69,7 +69,7 @@ line-length = 130
 target-version = ['py38']
 
 [tool.isort]
-src_paths = ["src/", "tests/"]
+src_paths = ["pyiceberg/", "tests/"]
 multi_line_output = 3
 profile = 'black'
 line_length = 130
